fix: problemIds 배열이 아닌 형태로 넘겨주는 곳 처리 및 에러 핸들링

+++ store/userInfo.js
@@ -58,6 +58,7 @@ const requestAddProblem = async () => {
 
 const requestDeleteProblem = async problemIds => {
   try {
+    if (!(problemIds instanceof Array)) throw new TypeError('problemIds 는 배열이어야합니다.');
     const res = await axios({
       method: 'delete',
       url: `delete`,
